Refactoring of characteristics Each Rule has his processor CharacPointProcessor is implemented Dices processors need some work (dice securisation)
