+++ src/DTO/Inputs/Characteristics/PointInput.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\DTO\Inputs\Characteristics;
+
+
+class PointInput extends InputAbstract
+{
+}
