+++ src/Domain/Model/test.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\Model;
+
+class test
+{
+
+
+}
+
