--- src/Domain/Personnages/Characteristiques/CharacRules.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Personnages\Characteristiques;
-
-enum CharacRules: string
-{
-    case Point = "point";
-    case Dice = 'dice';
-    case ThreeDices = 'three dices';
-}
