+++ src/Domain/Personnages/Characteristiques/Processors/CharacBuilderInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Domain\Personnages\Characteristiques\Processors;
+
+use App\Domain\Model\Caracteristique;
+use App\DTO\Inputs\Characteristics\InputAbstract;
+
+interface CharacBuilderInterface
+{
+    public function build(InputAbstract $baseCharac): Caracteristique;
+}
