+++ src/Domain/Personnages/Characteristiques/Processors/CharacProcessorAbstract.php
@@ -2,23 +2,22 @@
 
 declare(strict_types=1);
 
-namespace App\DTO;
+namespace App\Domain\Personnages\Characteristiques\Processors;
 
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProcessorInterface;
 use ApiPlatform\Api\IriConverterInterface;
-use App\Domain\Personnages\Characteristiques\CharacBuilderInterface;
-use App\Repository\CaracteristiqueRepository;
+
 use Doctrine\ORM\EntityManagerInterface;
 
-class CharacProcessor implements ProcessorInterface
+abstract class CharacProcessorAbstract implements ProcessorInterface, CharacBuilderInterface
 {
-    public function __construct(private IriConverterInterface $iriConverter, private CharacBuilderInterface $characBuilderInterface, private EntityManagerInterface $entityManager)
+    public function __construct(private IriConverterInterface $iriConverter, private EntityManagerInterface $entityManager)
     {
     }
     public function process(mixed $data, Operation $operation, array $uriVariables = [], array $context = [])
     {
-        $charac = $this->characBuilderInterface->build($data);
+        $charac = $this->build($data);
         $charac->setPersonnage($this->iriConverter->getResourceFromIri($data->getPersonnage()));
         $this->entityManager->persist($charac);
         $this->entityManager->flush();
