+++ src/Domain/Personnages/Competences/CompetencePersonnageBuilder.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\Personnages\Competences;
+
+class CompetencePersonnageBuilder
+{
+
+}
