+++ src/Domain/Personnages/Competences/CompetencePersonnageBuilderInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain\Personnages\Competences;
+
+interface CompetencePersonnageBuilderInterface
+{
+
+}
