+++ tests/Domain/Characteristics/ThreeDiceProcessorTest.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Tests\Domain\Characteristics;
+
+use PHPUnit\Framework\TestCase;
+
+class ThreeDiceProcessorTest extends TestCase
+{
+    use characConstruct;
+    public function testSomething(): void
+    {
+        $this->assertTrue(true);
+    }
+}
